# perf: math: precompute & use square of precisionReuse instead of 2 repeated computations