Noticed in an audit of Quicksilver's code while examining the Quo* code, that there were
2 invocations of a multiplication with precisionReuse of the simplified form:

```go
a *= b
a *= b
```
but those can be folded into

```go
a *= (b * b)
```

of which b is computed at init time, thus (b * b) can also be computed
at init time, so the final form becomes

```go
a *= b_squared
```

While here, added benchmarks for the change.

The results of this change yield performance improvements
```shell
$ benchstat before.txt after.txt
name                    old time/op    new time/op    delta
LegacyQuo-8               43.1ns ± 1%    37.1ns ± 3%  -14.00%  (p=0.000 n=14+15)
LegacyQuoTruncateMut-8    35.2ns ± 1%    28.8ns ± 3%  -18.27%  (p=0.000 n=15+15)
LegacyQuoRoundupMut-8      287ns ± 2%     283ns ± 3%     ~     (p=0.057 n=15+15)

name                    old alloc/op   new alloc/op   delta
LegacyQuo-8                0.00B          0.00B          ~     (all equal)
LegacyQuoTruncateMut-8     0.00B          0.00B          ~     (all equal)
LegacyQuoRoundupMut-8      72.0B ± 0%     72.0B ± 0%     ~     (all equal)

name                    old allocs/op  new allocs/op  delta
LegacyQuo-8                 0.00           0.00          ~     (all equal)
LegacyQuoTruncateMut-8      0.00           0.00          ~     (all equal)
LegacyQuoRoundupMut-8       2.00 ± 0%      2.00 ± 0%     ~     (all equal)
```

Fixes #12793